Tests 11 de febrero

+++ app/Http/Controllers/API/CicloFormativoController.php
@@ -7,6 +7,7 @@ use App\Models\CicloFormativo;
 use Illuminate\Http\Request;
 use App\Http\Resources\CicloFormativoResource;
 use App\Models\FamiliaProfesional;
+use Illuminate\Support\Facades\Gate;
 
 class CicloFormativoController extends Controller
 {
@@ -16,8 +17,8 @@ class CicloFormativoController extends Controller
     public function index(Request $request, FamiliaProfesional $familiaProfesional)
     {
         $query = CicloFormativo::where('familia_profesional_id', $familiaProfesional->id);
-        if ($query) {
-            $query->where('nombre', 'like', '%' . $request->q . '%');
+        if ($request->has('search')) {
+            $query->where('nombre', 'like', '%' . $request->search . '%');
         }
 
         return CicloFormativoResource::collection(
@@ -29,10 +30,21 @@ class CicloFormativoController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(Request $request, FamiliaProfesional $familiaProfesional)
     {
+
+
+        $request->validate([
+            'nombre' => 'required|string|max:255',
+            'codigo' => 'required|unique:ciclos_formativos,codigo|string|max:50',
+            'grado' => 'required|string|in:BÁSICA,G.M.,G.S.,C.E. (G.M.),C.E. (G.S.),basico,medio,superior',
+        ]);
+        //abort_if ($request->user()->cannot('create', CicloFormativo::class), 403);
+
         $cicloFormativo = json_decode($request->getContent(), true);
 
+        $cicloFormativo['familia_profesional_id'] = $familiaProfesional->id;
+        abort_if($request->user()->cannot('create', CicloFormativo::class), 403);
         $cicloFormativo = CicloFormativo::create($cicloFormativo);
 
         return new CicloFormativoResource($cicloFormativo);
@@ -50,6 +62,7 @@ class CicloFormativoController extends Controller
      */
     public function update(Request $request, FamiliaProfesional $familiaProfesional, CicloFormativo $cicloFormativo)
     {
+        abort_if($request->user()->cannot('update', $cicloFormativo), 403);
         $cicloFormativoData = json_decode($request->getContent(), true);
         $cicloFormativo->update($cicloFormativoData);
 
@@ -59,11 +72,14 @@ class CicloFormativoController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(FamiliaProfesional $familiaProfesional, CicloFormativo $cicloFormativo)
+    public function destroy(Request $request, FamiliaProfesional $familiaProfesional, CicloFormativo $cicloFormativo)
     {
+        abort_if($request->user()->cannot('delete', $cicloFormativo), 403);
         try {
             $cicloFormativo->delete();
-            return response()->json(null, 204);
+            return response()->json([
+                'message' => 'CicloFormativo eliminado correctamente'
+            ], 200);
         } catch (\Exception $e) {
             return response()->json([
                 'message' => 'Error: ' . $e->getMessage()
